refactored cert pin
